refactoring, add some comments
